+++ VEngine/Code/Actors/Unit.h
@@ -29,7 +29,7 @@ struct Unit : Actor
 
 	Unit();
 	virtual void Start() override;
-	virtual void Tick(double deltaTime) override;
+	virtual void Tick(float deltaTime) override;
 	virtual Properties GetProps() override;
 	void MoveTo(GridNode* destinationNode);
 };
