+++ VEngine/Code/Components/IComponentSystem.h
@@ -11,7 +11,7 @@ struct IComponentSystem
 	std::string name;
 
 	virtual void Init() = 0;
-	virtual void Tick(double deltaTime) = 0;
+	virtual void Tick(float deltaTime) = 0;
 	virtual void Start() = 0;
 	virtual void Cleanup() = 0;
 	virtual std::vector<Component*> GetComponents() = 0;
